ui/backend/requirements.txt
Flask>=3.0.0
Flask-SocketIO>=5.3.6
Flask-CORS>=4.0.0
python-socketio>=5.10.0
eventlet>=0.37.0

SQLAlchemy>=2.0.40
Flask-SQLAlchemy>=3.1.1
alembic>=1.14.0

watchdog>=3.0.0
psutil>=5.9.6

python-dotenv>=1.0.0
